Fixing the build issues for Dockerfile
- Also check_connection tool that wasn't working because of invalid
  schema.
- Also adding package-lock.json so "npm ci" command work
