+++ package.json
@@ -1,39 +1,38 @@
 {
-    "name": "sumologic-mcp-server",
-    "version": "1.0.0",
-    "description": "Model Context Protocol (MCP) server for SumoLogic API",
-    "main": "dist/index.js",
-    "scripts": {
-      "build": "tsc",
-      "start": "node dist/index.js",
-      "dev": "ts-node src/index.ts",
-      "test": "jest",
-      "lint": "eslint . --ext .ts",
-      "prepare": "npm run build"
-    },
-    "keywords": [
-      "sumologic",
-      "mcp",
-      "model-context-protocol",
-      "ai",
-      "api"
-    ],
-    "author": "",
-    "license": "MIT",
-    "dependencies": {
-      "@modelcontextprotocol/sdk": "1.12.1",
-      "axios": "^1.6.0",
-      "dotenv": "^16.3.1"
-    },
-    "devDependencies": {
-      "@types/jest": "^29.5.8",
-      "@types/node": "^20.8.10",
-      "@typescript-eslint/eslint-plugin": "^6.9.1",
-      "@typescript-eslint/parser": "^6.9.1",
-      "eslint": "^8.52.0",
-      "jest": "^29.7.0",
-      "ts-jest": "^29.1.1",
-      "ts-node": "^10.9.1",
-      "typescript": "^5.2.2"
-    }
+  "name": "sumologic-mcp-server",
+  "version": "1.0.0",
+  "description": "Model Context Protocol (MCP) server for SumoLogic API",
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "tsc",
+    "start": "node dist/index.js",
+    "dev": "ts-node src/index.ts",
+    "test": "jest",
+    "lint": "eslint . --ext .ts"
+  },
+  "keywords": [
+    "sumologic",
+    "mcp",
+    "model-context-protocol",
+    "ai",
+    "api"
+  ],
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "@modelcontextprotocol/sdk": "1.12.1",
+    "axios": "^1.6.0",
+    "dotenv": "^16.3.1"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.8",
+    "@types/node": "^20.8.10",
+    "@typescript-eslint/eslint-plugin": "^6.9.1",
+    "@typescript-eslint/parser": "^6.9.1",
+    "eslint": "^8.52.0",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   }
+}
